+++ packages/veritone-widgets/src/widgets/MediaPlayer/index.js
@@ -83,6 +83,7 @@ class MediaPlayerComponent extends React.Component {
     isActive: bool,
     paused: bool,
     currentTime: number,
+    autofocus: bool,
     forwardedRef: objectOf(any)
   };
 
@@ -140,6 +141,7 @@ class MediaPlayerComponent extends React.Component {
             stagedBoundingBoxStyles={props.stagedBoundingBoxStyles}
             stylesByObjectType={props.stylesByObjectType}
             defaultBoundingBoxStyles={props.defaultBoundingBoxStyles}
+            autofocus={props.autofocus}
           />
         )}
         <Player